made changes in response to feedback
